level2 (structure): refactor subviews
